:bug: Fix #19

+++ src/main/java/org/b3log/solo/processor/console/StaticSiteConsole.java
@@ -15,23 +15,26 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package org.b3log.solo.processor;
+package org.b3log.solo.processor.console;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.lang.time.DateFormatUtils;
 import org.b3log.latke.Latkes;
+import org.b3log.latke.http.HttpMethod;
 import org.b3log.latke.http.RequestContext;
 import org.b3log.latke.http.annotation.Before;
 import org.b3log.latke.http.annotation.RequestProcessing;
 import org.b3log.latke.http.annotation.RequestProcessor;
 import org.b3log.latke.ioc.BeanManager;
+import org.b3log.latke.ioc.Inject;
 import org.b3log.latke.logging.Level;
 import org.b3log.latke.logging.Logger;
+import org.b3log.latke.service.LangPropsService;
 import org.b3log.latke.util.CollectionUtils;
+import org.b3log.latke.util.Strings;
 import org.b3log.solo.model.*;
-import org.b3log.solo.processor.console.ConsoleAuthAdvice;
 import org.b3log.solo.service.*;
 import org.b3log.solo.util.Mocks;
 import org.b3log.solo.util.Solos;
@@ -44,52 +47,50 @@ import java.nio.charset.StandardCharsets;
 import java.util.List;
 
 /**
- * Static site processor. HTML 静态站点生成 https://github.com/88250/solo/issues/19
+ * Static site console request processing. HTML 静态站点生成 https://github.com/88250/solo/issues/19
  *
  * @author <a href="http://88250.b3log.org">Liang Ding</a>
- * @version 1.0.0.1, Jan 8, 2020
+ * @version 1.0.0.0, Jan 13, 2020
  * @since 3.9.0
  */
 @RequestProcessor
-@Before(ConsoleAuthAdvice.class)
-public class StaticSiteProcessor {
+@Before(ConsoleAdminAuthAdvice.class)
+public class StaticSiteConsole {
 
     /**
      * Logger.
      */
-    private static final Logger LOGGER = Logger.getLogger(StaticSiteProcessor.class);
+    private static final Logger LOGGER = Logger.getLogger(StaticSiteConsole.class);
 
     /**
-     * Name of generate directory.
+     * Language service.
      */
-    private static final String STATIC_SITE = "static-site";
-
-    /**
-     * Path of generate directory.
-     */
-    private static final String staticSitePath = StaticSiteProcessor.class.getResource("/" + STATIC_SITE).getPath();
-
-    /**
-     * Source directory path.
-     */
-    private static final String sourcePath = StaticSiteProcessor.class.getResource("/").getPath();
+    @Inject
+    private LangPropsService langPropsService;
 
     /**
      * Generates static site.
      *
-     * @param context the specified context
+     * @param context the specified request context
      */
-    @RequestProcessing(value = "/static-site")
-    public synchronized void genStaticSite(final RequestContext context) {
+    @RequestProcessing(value = "/console/staticsite", method = HttpMethod.PUT)
+    public synchronized void genSite(final RequestContext context) {
         try {
+            final JSONObject requestJSONObject = context.requestJSON();
+            final String url = requestJSONObject.optString(Common.URL);
+            if (!Strings.isURL(url)) {
+                context.renderJSON(-1);
+                context.renderMsg("Invalid site URL");
+
+                return;
+            }
+
             FileUtils.deleteDirectory(new File(staticSitePath));
             FileUtils.forceMkdir(new File(staticSitePath));
 
             // 切换至静态站点生成模式
             Latkes.setServerScheme("https");
-            // TODO: 前端传入生成站点域名
-            Latkes.setServerHost("88250.github.io");
-//            Latkes.setServerHost("dl88250.gitee.io");
+            Latkes.setServerHost(url);
             Latkes.setServerPort("");
             Solos.GEN_STATIC_SITE = true;
 
@@ -121,12 +122,33 @@ public class StaticSiteProcessor {
             Solos.GEN_STATIC_SITE = false;
 
             LOGGER.log(Level.INFO, "Static site generated [dir=" + staticSitePath + "]");
+
+            String siteGenedLabel = langPropsService.get("siteGenedLabel");
+            siteGenedLabel = siteGenedLabel.replace("{dir}", staticSitePath);
             context.renderJSON(0);
+            context.renderMsg(siteGenedLabel);
         } catch (final Exception e) {
             LOGGER.log(Level.ERROR, "Generates static site failed", e);
+
             context.renderJSON(-1);
+            context.renderMsg(langPropsService.get("updateFailLabel"));
         }
     }
+
+    /**
+     * Name of generate directory.
+     */
+    private static final String STATIC_SITE = "static-site";
+
+    /**
+     * Path of generate directory.
+     */
+    private static final String staticSitePath = StaticSiteConsole.class.getResource("/" + STATIC_SITE).getPath();
+
+    /**
+     * Source directory path.
+     */
+    private static final String sourcePath = StaticSiteConsole.class.getResource("/").getPath();
 
     private static void genCategories() throws Exception {
         final BeanManager beanManager = BeanManager.getInstance();
@@ -248,28 +270,28 @@ public class StaticSiteProcessor {
     private static void genSkins() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/skins"));
         FileUtils.forceMkdir(new File(staticSitePath + "/skins"));
-        FileUtils.copyDirectory(new File(StaticSiteProcessor.class.getResource("/skins").toURI()), new File(staticSitePath + "/skins"));
+        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/skins").toURI()), new File(staticSitePath + "/skins"));
         LOGGER.log(Level.INFO, "Generated skins");
     }
 
     private static void genJS() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/js"));
         FileUtils.forceMkdir(new File(staticSitePath + "/js"));
-        FileUtils.copyDirectory(new File(StaticSiteProcessor.class.getResource("/js").toURI()), new File(staticSitePath + "/js"));
+        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/js").toURI()), new File(staticSitePath + "/js"));
         LOGGER.log(Level.INFO, "Generated js");
     }
 
     private static void genImages() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/images"));
         FileUtils.forceMkdir(new File(staticSitePath + "/images"));
-        FileUtils.copyDirectory(new File(StaticSiteProcessor.class.getResource("/images").toURI()), new File(staticSitePath + "/images"));
+        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/images").toURI()), new File(staticSitePath + "/images"));
         LOGGER.log(Level.INFO, "Generated images");
     }
 
     private static void genPlugins() throws Exception {
         FileUtils.deleteDirectory(new File(staticSitePath + "/plugins"));
         FileUtils.forceMkdir(new File(staticSitePath + "/plugins"));
-        FileUtils.copyDirectory(new File(StaticSiteProcessor.class.getResource("/plugins").toURI()), new File(staticSitePath + "/plugins"));
+        FileUtils.copyDirectory(new File(StaticSiteConsole.class.getResource("/plugins").toURI()), new File(staticSitePath + "/plugins"));
         genURI("/plugins/kanbanniang/assets/model.json");
         LOGGER.log(Level.INFO, "Generated plugins");
 
@@ -277,8 +299,9 @@ public class StaticSiteProcessor {
 
     private static void genFile(final String file) throws Exception {
         FileUtils.forceMkdirParent(new File(staticSitePath + "/" + file));
-        final String staticSitePath = StaticSiteProcessor.class.getResource("/" + STATIC_SITE).toURI().getPath();
+        final String staticSitePath = StaticSiteConsole.class.getResource("/" + STATIC_SITE).toURI().getPath();
         FileUtils.copyFile(new File(sourcePath + "/" + file), new File(staticSitePath + "/" + file));
         LOGGER.log(Level.INFO, "Generated a file [" + file + "]");
     }
+
 }
